Separate test class / add tests for invalid image URL
